Refactor: relocate SDMA completion detail into backend/sdma/ (PR-B)
Pure file move + include adjustment, no behavior change.

- New backend/sdma/sdma_completion_scheduler.h holds the SdmaEventRecord
  mirror + raw-addr poll/retire helpers; pto_async_wait.h now includes it
  and keeps only thin one-line ops-table glue.
- New backend/sdma/sdma_completion_kernel.h holds defer_pto_async_event +
  defer_sdma_event_record + the PTO-ISA SDMA-detail includes; the example
  picks them up from there.
- CompletionPollState / CompletionPollResult move to pto_completion_token.h
  so the SDMA scheduler header doesn't have to pull in pto_async_wait.h.
- pto_async_kernel_api.h no longer carries any SDMA-specific helper.

Verified: a2a3-onboard runtime builds; sdma_async_completion_demo,
async_notify_demo, deferred_notify_demo all pass with max_diff=0.
