+++ examples/a2a3/tensormap_and_ringbuffer/sdma_async_completion_demo/kernels/aiv/kernel_sdma_tget_async.cpp
@@ -22,6 +22,7 @@
 #include <pto/npu/comm/async/sdma/sdma_types.hpp>
 #include <pto/pto-inst.hpp>
 
+#include "backend/sdma/sdma_completion_kernel.h"
 #include "platform_comm/comm_context.h"
 #include "pto_async_kernel_api.h"
 #include "tensor.h"
